requirements.txt
typer>=0.9.0
pydantic>=2.5.0
pandas>=2.1.0
polars>=0.19.0
rich>=13.0.0
networkx>=3.1
shapely>=2.0.0
geopandas>=0.14.0
requests>=2.31.0
pytest>=6.0
ruff>=0.1.0
folium>=0.14.0
jinja2>=3.1.2
openai>=1.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
